requirements.txt
django<5
djangorestframework
django-cors-headers